Improved spaces support
Add a route that shows all spaces.

Add a route that shows all changes (over all spaces).

Add tests for both.

Add a "dev" make target that uses morbo (which comes with
Mojolicious), a watcher that starts the server and restarts it
whenever gemini-wiki.pl changes.

Remove a call to run() which looks superfluous.

Remove an occurence of $space in a log message that would produce a
warning when it was undefined.