Add virtual hosting
This required a lot of changes. Most importantly, the --host option
specifies which requests the Gemini Wiki serves. The requests includes
the hostname; it is available via Gemini URL or Host field in the HTTP
headers. We check that the hostname matches one of the hosts the
Gemini Wiki serves, and if it does, we use this hostname for all the
links. This means, passing the hostname to all the functions...