Allow working with client certificates
The Makefile includes two new targets: client-cert makes a new
certificate for a client. client-fingerprint shows a fingerprint that
is suitable for use with the IO::Socket::SSL library that Gemini Wiki
uses. The code is added to the Makefile in order to have an executable
example available. I don't expect administrators to create client
certificates for their users!

The documentation now includes an example of how to write a config
file that prevents the editing of the wiki except for known
fingerprints.