Fix a warning in read_log
File::ReadBackwards->readline sometimes returns an undefined value and
sometimes a newline. I'm not quite sure why I'm getting this newlines
every now and then. The problem is that such lines are defined but
don't split into fields, resulting in $host being used as $ts,
resulting in a warning when these records get sorted as
'alexschroeder.ch' is not numeric.