Fix handling of files in changes
When two namespaces have a file with the same $id, then the second
occurence of the file is not "the same file" (has not been %seen).
Which is why %seen must contain the $name and not the $id. The $name
contains $host and $space in addition to $id, if necessary.