Use write_text when deleting a page
The previous code of sub delete used open and all that, which is fine,
but long. But then it also used say $fh @pages which doesn't separate
the items with a newline, apparently? Or at least not in this case.
Just to be sure, we're now using the same code as in sub pages:

write_text($index, join("\n", @pages, ""));