Sanitize .pem handling
When the server starts up and doesn't have a cert.pem and key.pem
files (and the file names weren't changed from these defaults), it
prompts the user for a hostname and generates them for you.

When the test server starts up and doesn't have a cert.pem and key.pem
files in the t directory, it generates them for you.