Increase version to 1.22
The archive tests are skipped if /bin/tar doesn't exist, and the
openssl version used is reported when generating a new certificate.
With these two changes I'm hoping to improve the test reports for BSD
and to gain new insight into the test environment for the old
"3.10.0-1127.10.1.el7.x86_64" linux environment (RHEL17).