Fix test setup for RHEL7
Apparently the problem is that RHEL7 ships with OpenSSL 1.0.2k which
does Elliptic Curves differently. The OpenSSL wiki page says the
following:

    If you use a key or certificate without without the
    OPENSSL_EC_NAMED_CURVE flag (i.e., one that looks like the image
    on the right), then the SSL connection will fail with the
    following symptoms: … Server … 140339533272744:error:1408A0C1:SSL
    routines:SSL3_GET_CLIENT_HELLO:no shared cipher:s3_srvr.c:1353:

For reference, what I’m seeing in the RHEL7 tests on CPAN is this:

    1408A0C1:SSL routines:ssl3_get_client_hello:no shared cipher

Thus, I'm now checking the OpenSSL version before generating keys and
when OpenSSL 1.0 is used, RSA keys are generated.

https://wiki.openssl.org/index.php/Elliptic_Curve_Cryptography