Move from Net::Server to Mojo::IOLoop
Use Mojo::IOLoop. This is the big one!

The functions on @extensions are called using a different first
parameter. Instead of it being $self, the Net::Server, it's now
$stream, a Mojo::IOLoop::Stream. Instead of using say or print to send
output, you now have to use $stream->write(). All the other functions
you used to call now aren't part of $self anymore but simply in the
App::Phoebe package. I know, it's a pain. I'm sorry. the benefit is
that we now have a streaming server.

@init was removed. Use the config file to set options.

--log_level now takes error, warn, info, debug instead of numbers.

No longer provide many of the options Net::Server provided. That
means, all of these no longer work:

--setsid
--pid_file
--log_file
--user
--group

Use systemd intead.