Add web-edit, make appropriate changes to Phoebe
Web-edit produces a form to edit pages via the web, and handles POST
requests to save them.

In order for this to work, Phoebe needs pass along the buffer for all
web requests having a Content-Length header, since these could be PUT
or POST requests. The run_extensions function will pass it along to
extensions such as the extension used by the web-edit code.

As it turns out the space_regex function doesn't actually need the
$stream parameter so it was removed.

Add an error if process_http doesn't actually handle the HTTP request.
Close the $stream at the end of process_http such that the end of
process_http and process_gemini are the same.