Make test.pl wait up to 5s
As CPAN Testers on Windows kept reporting a failure, I tried it myself
on a Windows machine and at least on the one I was using the problem
was that 1. the firewall was blocking Perl, requiring an administrator
password; 2. IO::Socket::SSL did not install due to errors unless I
installed it via Cygwin; 3. the client needs to wait for a few
seconds (more than 2s but 5s or less) for the server to start up
before testing can begin. So now we just check every second for five
times when loading t/test.pl.