Close some streams on errors
Also switch the test client to Mojo::IOLoop, i.e. a streaming server.
This means that it returns Unicode strings instead of normal strings,
so some other tests had to be updated as well. The main benefit,
however, is that with the streaming server, and the closed streams, a
test in t/basics.t could be added back: testing for the wrong MIME
type.