Change the lock test to use Mojo::IOLoop->timer
The code using alarm worked well for me, but suddenly I'm getting a
lot of errors from CPAN testers and it makes me think that perhaps the
alarm(1) call sometimes more than 1s to go off, and if it takes more
than 5s, the test fails. So let's try the timer from the same event
loop that Phoebe is using.