Attempt to fix the memory leak
The IO::Socket::SSL man page says: "If you have a server and it looks
like you have a memory leak you might check the size of your session
cache. Default for Net::SSLeay seems to be 20480, see the example for
SSL_create_ctx_callback for how to limit it."

OK. I'm going to try that instead. The previous code had no effect.