Windows is declared unsupported
CPAN Testers report that the tests hang on Windows. I don't know why
this is so, but I also don't think many people are running web servers
and the like on their Windows machines, so this seems appropriate.

If you are one of the few who wants to try it, I think there's nothing
fundamentally incompatible with Windows, I just don't know what the
problem is. Patches welcome.