oddmuse: add support for namespace links
This adds more rules: numbered entities are not hashtags, e.g.
&#x2605; is ★; free links without text, e.g. [[Alex Schröder]]; free
links with text, e.g. [[Alex Schroeder|code monkey]]; namespaces, e.g.
on campaignwiki.org: [Tau_Subsector:HomePage Tau Subsector]; but
handle the index action, e.g. [Tau_Subsector:?action=index 39 pages];
wiki words with text, e.g. [AlexSchroeder code monkey].

Phoebe only starts the servers if and only if the script that's
running ($0) is the same as the name of the script (__FILE__, i.e.
"phoebe"). When the script is loaded as a library by some other script
this isn't true, and the servers aren't started. This allows us to run
unit tests in t/oddmuse.t that don't actually need the server, and
actual pages being saved.

For unit tests to work, the port sub needs to return a number even if
$stream is undefined.

And finally, Phoebe needs to return a positive value when it's loaded
as a library.