Hand web editing via the iPhone
Currently, the iPhone sends the POST request in two parts. The first
read will get us all the headers, the second read will get us the
message body. This used to cause a problem, because the code assumed
that once we got to the body, we got it all in one swoop (instead of
the 0 bytes); the code would then proceed to parse it, find no token,
and print an error. It would also continue reading the rest of the
HTTP request, find the rest of the body, parse it, and save. The
problem was that the previous error message ruined the HTTP response.

Now HTTP body parsing does the same loop as HTTP header parsing: if
the data isn't complete, we wait until it is.