More changes to ijirait
Wrap the notification code for streamers in an eval, because it seems
to cause errors every now and then. On the other hand, ever since
making this change, the error could not be reproduced, so currently
it's just there, attempting to log problems. We'll see.

Make sure to return 1 when handling a command that doesn't require a
cert. Without this change, getting the RSS feed would get you the RSS
feed plus a bogus header at the end saying that your request cannot be
handled. Return 1 tells Phoebe that we did in fact handle the request.

Change the RSS from reporting the things said (which feels a bit
creepy) to the last activity of every character, which is similar to
the who command.