Improved Oddmuse/Gopher compatibility
Moving the newest_first sort into the pages sub so that we can skip
the sorting if and only if we're requesting the page list from
Oddmuse.

Fix some Gopher bugs in the menu structure.

Fix the Oddmuse spaces regex if no spaces are defined.

Fix some issues with calls using @_: can't use @_ in a call if we used
shift to get the arguments as those arguments are no longer in @_.