Image links leave no inline text
Thus, a paragraph containing a link like [[image:this]] ends up as
"Thus, a paragraph containing a link like  ends up as" ← that is, the
word "this" is missing. The reasoning is that such links are always
shown outside regular flow in a HTML document: either on a paragraph
of their own, or floating to the left or to the right.