App::Phoebe::RegisteredEditorsOnly for the web
Make sure RegisteredEditorsOnly works for the web, too. Add an example
of how to generate the PKCS12 file you need to add the client cert to
Firefox. Write a test that shows how the example config works.

Fix App::Phoebe::WebEdit such that it no longer requires a token on
the form if wiki tokens is an empty list.