Improved handling of data using gemini
The gemini script now checks whether the tty it is connected to (if
any) can display encoded data. It refuses to print mismatched
encodings (i.e. if your terminal cannot display UTF-8) or binary
data (i.e. images) unless the output is redirected into a file or a
pipe, and it says as much. Use --force to override this.

Also fix a bug involving double encoding of path or query that already
contained percent-encoded characters.