Don't close the stream when with_lock succeeds
When running code in with_lock, and it works, there's no need to close
the stream. In Phoebe.pm, this is not a problem, since with_lock
comes at the end of save_page, which comes at the end of
process_titan, but in the new Oracle.pm, this prevented the printing
of the result.