Close the stream after saving a page
Actually, we want the stream to remain open if code gets called
with_lock, but in the case of save_page, we want the stream to be
closed if it successfully saved the page. If the code is has to wait
for the lock, we aren't finished with the stream, yet, so we can't
close the stream after calling with_lock!

There still is a problem with extensions, though. If they reschedule,
process_gemini will have closed the stream, I fear.