Fixed the read_lines problem
undef $/ has global effect, and is rarely called as part of the Oracle
example code to notify Antenna of new questions. When I had it in
production, that was the time when read_lines started failing. I no
longer use read_lines, but it was still worth fixing this bug.