Remove host and port from web links
This makes it possible to run the site behind a reverse proxy on the
web. The web server then uses --no_cert to proxy the connection to
Phoebe, which obviously is not listening on port 443 but on some other
port. Therefore the local links must not include that other port…